Add sigils for quiet and allow for arbitary prefixes.
Previously, only `quiet op` was allowed. Now one can do `# op` or `♯ op`  for the same functionality.

A dictionary of prefixes have been added. This allows for more prefixes or sigils if we want.

I choose to parse prefixes in `parse_program` and not in `Instr` because Instr should be a simple wrapper object. It isn't responsible for parsing instructions.

Ideally, this also means that `sigil_to_op` should be moved to `parse_program`, but that will come in a later commit.